진화 관련 메소드 파일 분리
